backend/requirements.txt
# Core
fastapi==0.128.0
starlette==0.50.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1

# Phase III - AI Chatbot (Ollama + LangChain + MCP)
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.1
mcp==1.25.0
httpx==0.27.2
aiofiles==24.1.0
